Rename UI package pcgen.gui2.doomsdaybook to pcgen.gui2.namegen
Mirrors the engine package rename. NameGenPanel, NameButton, and
XMLFilter move; RandomNameDialog updates its single import.
