+++ code/src/java/pcgen/gui2/namegen/NameGenPanel.java
@@ -15,7 +15,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
-package pcgen.gui2.doomsdaybook;
+package pcgen.gui2.namegen;
 
 import java.awt.BorderLayout;
 import java.awt.FlowLayout;
